Move addresses into separate file
